requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
jupyter==1.0.0
ipykernel==6.27.1
